% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogramDistance.R
\name{histogramDistance}
\alias{histogramDistance}
\title{Plot two StochKit2R simulation histograms and display distance}
\usage{
histogramDistance(histogramData1, histogramData2, file1 = FALSE,
  file2 = FALSE)
}
\arguments{
\item{histogramData1}{histogram data from output object or string with path to StochKit2 histogram output file 1. IMPORTANT: histogram file names have format hist_<species index>_<time point>.dat, where species index STARTS AT 0 (not 1!)}

\item{histogramData2}{histogram data from output object or string with path to StochKit2 histogram output file 2. IMPORTANT: histogram file names have format hist_<species index>_<time point>.dat, where species index STARTS AT 0 (not 1!)}

\item{file1}{set to TRUE if \code{histogramData1} is a path to the histogram output file (rather than returned output data)}

\item{file2}{set to TRUE if \code{histogramData2} is a path to the histogram output file (rather than returned output data)}
}
\value{
The ggplot object
}
\description{
\code{histogramDistance} Plots histograms of data stored in StochKit2R histogram returned output or output files and calculates histogram distance. IMPORTANT: histogram file names have format hist_<species index>_<time point>.dat, where species index STARTS AT 0 (not 1!)
}
\examples{
\dontrun{
#example using included dimer_decay.xml file
model <- system.file("dimer_decay.xml",package="StochKit2R")
#output written to ex_out directory (created in current working directory)
result1 = ssa(model,10,100,20,keepHistograms=TRUE,outputDir="ex_out",force=TRUE)

#another ensemble with output written to ex_out2 directory
result2 = ssa(model,10,100,20,keepHistograms=TRUE,outputDir="ex_out",force=TRUE)
#plot the histograms for species 2 ("S2") at time point 5 (t=2.0) for the two runs above
histogramDistance(result1$hist$S2[[5]],result2$hist$S2[[5]])
#IMPORTANT: histogram file names have format:
#hist_<species index>_<time point>.dat, where species index STARTS AT 0
histogramDistance("ex_out/histograms/hist_1_4.dat",
                 "ex_out2/histograms/hist_1_4.dat",file1=TRUE,file2=TRUE)
}
}
