% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ode.R
\name{ode}
\alias{ode}
\title{Gillespie Stochastic Simulation Algorithm ordinary differential equation simulation}
\usage{
ode(modelFile, time, intervals)
}
\arguments{
\item{modelFile}{Character string with path to StochKit2 .xml model file}

\item{time}{Simulation time}

\item{intervals}{Number of output intervals. Default 0 outputs at end time only. 1=keep data at start and end time, 2=keep data at start, middle, and end times, etc. Note data is stored at (intervals+1) equally spaced time points.}
}
\value{
Data frame containing time and species population
}
\description{
\code{ode} Run an ODE simulation (population version of the reaction rate equations). Homodimerization (e.g. A+A type) propensities are converted to Reaction Rate Equation rates (c*A*A instead of c*A*(A-1)/2) for mass action reactions. Propensities of customized propensities are not converted. Uses the 5th order Dormand-Prince Runge-Kutta method.
}
\examples{
\dontrun{
#example using included dimer_decay.xml file
out <- ode(system.file("dimer_decay.xml",package="StochKit2R"),time=10,intervals=20)
}
}
