% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{odecpp}
\alias{odecpp}
\title{C++ Interface to Gillespie Stochastic Simulation Algorithm single trajectory}
\usage{
odecpp(StochKit2Rmodel, endTime, intervals)
}
\arguments{
\item{StochKit2Rmodel}{R list (Rcpp List built from buildStochKit2Rmodel output)}

\item{endTime}{Simulation end time}
}
\value{
Dataframe containing the time and population sizes
}
\description{
\code{ode} Called by StochKit2R ode function, do not call this C++ interface directly
}
\keyword{internal}
