% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStats.R
\name{plotStats}
\alias{plotStats}
\title{Plot StochKit2R simulation statistics output data}
\usage{
plotStats(statsDF, indices)
}
\arguments{
\item{statsDF}{ssa output, list of data frames}

\item{indices}{The species indices that will be plotted. The first species is index 1}
}
\value{
The ggplot object
}
\description{
\code{plotStats} Plots means and mean +/- one standard deviation of populations specified in \code{indices} in the StochKit2R stats output directory \code{statsDirectory}
}
\examples{
\dontrun{
#example using included dimer_decay.xml file
model <- system.file("dimer_decay.xml",package="StochKit2R")
#output written to ex_out directory (created in current working directory)
out <- ssa(model,"ex_out",10,100,20,force=TRUE)
#plot the data for species 1,2 and 3 (all of them in the dimer decay model)
plotStats(out,c(1,2,3))
}
}
