% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrajectories.R
\name{plotTrajectories}
\alias{plotTrajectories}
\title{Plot StochKit2R simulation trajectory output data}
\usage{
plotTrajectories(trajectoriesData, outputIndexes, speciesIndexes, file = F)
}
\arguments{
\item{trajectoriesData}{trajs list from ssa (or tauLeaping) returned output or character string containing the path to the trajectories output directory}

\item{outputIndexes}{Integer vector containing indexes for which trials will be plotted}

\item{speciesIndexes}{Integer vector of the species indices that will be plotted. The first species is index 1}

\item{file}{set to TRUE if trajectoriesData is a path to the trajectories output directory (rather than returned output data)}
}
\value{
The ggplot object
}
\description{
\code{plotTrajectories} Plots trajectories \code{outputIndexes} of populations specified in \code{speciesIndexes} in the StochKit2R trajectories in \code{trajectoriesData}
}
\examples{
\dontrun{
#example using included dimer_decay.xml file
model <- system.file("dimer_decay.xml",package="StochKit2R")
#output written to ex_out directory (created in current working directory)
out <- ssa(model,time=10,realizations=100,intervals=20,keepTrajectories=TRUE,outputDir="ex_out",force=TRUE)
#plot the data for species 1,2 and 3 for trajectories 2,3,4 and 5 from file
plotTrajectories("ex_out/trajectories",2:3,1:3,file=TRUE)
#same plot from output
plotTrajectories(out$trajs,2:3,1:3)
}
}
