% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEnsembleData.R
\name{readEnsembleData}
\alias{readEnsembleData}
\title{Read StochKit2R ensemble output files (reads data from files written by writeEnsembleData)}
\usage{
readEnsembleData(outputDir)
}
\arguments{
\item{outputDir}{Character string with path to output directory.}
}
\description{
\code{writeEnsembleOutput} Reads StochKit2R ensemble output data from directory. If data is from StochKit2R C++ starting indexes will be converted from 0 to 1.
}
\examples{
\dontrun{
#'#example using included dimer_decay.xml file
#run 100 simulations for 10 time units, keeping output at 20 time intervals
#store model file name in a variable first
model <- system.file("dimer_decay.xml",package="StochKit2R")
out <- ssa(modelFile=model,time=10,realizations=100,intervals=20)
#creates ex_out directory and writes output
writeEnsembleData(data=out,outputDir="ex_out")
out2=readEnsembleData("ex_out") #out and out2 should match
}
}
