% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssa.R
\name{ssa}
\alias{ssa}
\title{Gillespie Stochastic Simulation Algorithm}
\usage{
ssa(
  modelFile,
  time,
  realizations,
  intervals = 0,
  noStats = FALSE,
  keepTrajectories = FALSE,
  keepHistograms = FALSE,
  bins = 32,
  seed = NULL,
  p = 0
)
}
\arguments{
\item{modelFile}{Character string with path to StochKit2 .xml model file}

\item{time}{Simulation time of each realization}

\item{realizations}{Number of realizations}

\item{intervals}{Number of output intervals. Default 0 outputs at end time only. 1=keep data at start and end time, 2=keep data at start, middle, and end times, etc. Note data is stored at (intervals+1) equally spaced time points.}

\item{noStats}{Do not keep means and variances data. Default \code{FALSE} creates stats output.}

\item{keepTrajectories}{Keep trajectory data.}

\item{keepHistograms}{Keep histogram data.}

\item{bins}{Number of histogram bins}

\item{seed}{Seed the random number generator. By default the seed is determined by the R random number generator, so the seed can also be set by calling \code{set.seed} in R immediately before calling \code{ssa}}

\item{p}{Override default and specify the number of processes (threads) to use. By default (=0), the number of processes will be determined automatically (recommended). Ignored on systems without OpenMP support.}
}
\value{
List of statistics, trajectory and histogram output data.
}
\description{
\code{ssa} Runs an ensemble of SSA \code{realizations} (trajectories)
}
\examples{
\dontrun{
#'#example using included dimer_decay.xml file
#run 100 simulations for 10 time units, keeping output at 20 time intervals
#store model file name in a variable first
model <- system.file("dimer_decay.xml", package="StochKit2R")
out <- ssa(modelFile=model, time=10, realizations=100, intervals=20)
#more typical example where model file is stored elsewhere
#(must be valid path to existing .xml StochKit2 model file)
#also, keep trajectory data.
out <- ssa("Desktop/dimer_decay.xml", 10, 100, 20, keepTrajectories=TRUE)
}
}
