% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ssaSingle.R
\name{ssaSingle}
\alias{ssaSingle}
\title{Gillespie Stochastic Simulation Algorithm single trajectory}
\usage{
ssaSingle(modelFile, time, seed = NULL)
}
\arguments{
\item{modelFile}{Character string with path to StochKit2 .xml model file}

\item{time}{Simulation end time}

\item{seed}{Seed the random number generator. By default the seed is determined by the R random number generator, so the seed can also be set by calling \code{set.seed} in R immediately before calling \code{ssaSingle}}
}
\value{
Data frame containing time and species population for every reaction event in the simulation.
}
\description{
\code{ssaSingle} Run a single SSA trajectory. Stores a row of data for every reaction event.
}
\examples{
\dontrun{
#example using included dimer_decay.xml file
out <- ssaSingle(system.file("dimer_decay.xml",package="StochKit2R"),time=10)
}
}
