% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ssaSingleStochKit2RInterface}
\alias{ssaSingleStochKit2RInterface}
\title{C++ Interface to Gillespie Stochastic Simulation Algorithm single trajectory}
\usage{
ssaSingleStochKit2RInterface(
  StochKit2Rmodel,
  startTime,
  endTime,
  outputFileNameString,
  seed
)
}
\arguments{
\item{StochKit2Rmodel}{R list (Rcpp List built from buildStochKit2Rmodel output)}

\item{startTime}{Simulation start time}

\item{endTime}{Simulation end time}

\item{outputFileNameString}{Character string with path to output file.}

\item{seed}{Seed for random number generator}
}
\value{
Dataframe containing the time and population sizes
}
\description{
\code{ssa} Called by StochKit2R ssaSingle function, do not call this C++ interface directly
}
\keyword{internal}
