% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StochKit2R.R
\docType{package}
\name{StochKit2R}
\alias{StochKit2R-package}
\alias{StochKit2R}
\title{StochKit2R}
\description{
StochKit2R: a package for efficient discrete stochastic simulation
using the Gillespie Algorithm (SSA) and tau-leaping. The package
provides an R interface to stochastic simulation functions and
provides functions for generating plots of the simulation data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/kevinsanft/StochKit2R}
  \item \url{http://www.kevinsanft.com/StochKit2R}
  \item Report bugs at \url{https://github.com/kevinsanft/StochKit2R/issues}
}

}
\author{
\strong{Maintainer}: Kevin R. Sanft \email{kevin@kevinsanft.com} [copyright holder]

Other contributors:
\itemize{
  \item Harris Joseph [contributor]
  \item John Ravi [contributor]
  \item Linda R. Petzold [copyright holder]
}

}
