% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tauLeaping.R
\name{tauLeaping}
\alias{tauLeaping}
\title{Explicit Adaptive Tau-leaping simulation}
\usage{
tauLeaping(modelFile, time, realizations, intervals = 0, noStats = FALSE,
  keepTrajectories = FALSE, keepHistograms = FALSE, bins = 32,
  outputDir = NULL, force = FALSE, seed = NULL, p = 0, epsilon = 0.03,
  threshold = 10)
}
\arguments{
\item{modelFile}{Character string with path to StochKit2 .xml model file}

\item{time}{Simulation time of each realization}

\item{realizations}{Number of realizations}

\item{intervals}{Number of output intervals. Default 0 outputs at end time only. 1=keep data at start and end time, 2=keep data at start, middle, and end times, etc. Note data is stored at (intervals+1) equally spaced time points.}

\item{noStats}{Do not keep means and variances data. Default \code{FALSE} creates stats output.}

\item{keepTrajectories}{Keep trajectory data.}

\item{keepHistograms}{Keep histogram data.}

\item{bins}{Number of histogram bins}

\item{outputDir}{Character string with path to output directory. By default (=NULL) no data is written to file. If specified output directory does not exist, it will be created. If output directory already exists, use \code{force=TRUE} to overwrite}

\item{force}{Force overwriting of existing output directory data. No effect when \code{outputDir=NULL}.}

\item{seed}{Seed the random number generator. By default the seed is determined by the R random number generator, so the seed can also be set by calling \code{set.seed} in R immediately before calling \code{tauLeaping}}

\item{p}{Override default and specify the number of processes (threads) to use. By default (=0), the number of processes will be determined automatically (recommended). Ignored on systems without OpenMP support.}

\item{epsilon}{Set the tolerance (applicable to tauLeaping only), default is 0.03. Valid values: must be greater than 0.0 and less than 1.0}

\item{threshold}{Set the threshold (minimum number of reactions per leap before switching to ssa) for tauLeaping}
}
\value{
List of data frames with statistics, trajectory and histogram output data.
}
\description{
\code{tauLeaping} Runs an ensemble of Tau-leaping \code{realizations} (trajectories). Will revert to SSA if Tau-leaping is not beneficial.
}
\examples{
\dontrun{
#example using included dimer_decay.xml file
#store path to StochKit2R dimer_decay.xml file in a variable
model <- system.file("dimer_decay.xml",package="StochKit2R")
out <- tauLeaping(modelFile=model,time=10,realizations=100,intervals=20)

#more typical example where model file is stored elsewhere
#(must be valid path to existing .xml StochKit2 model file)
#also, keep trajectory data
out <- tauLeaping("Desktop/dimer_decay.xml",10,100,20,keepTrajectories=TRUE)
}
}
