% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeEnsembleData.R
\name{writeEnsembleData}
\alias{writeEnsembleData}
\title{Write StochKit2R data to files}
\usage{
writeEnsembleData(data, outputDir, force = FALSE, indexStart = 1)
}
\arguments{
\item{data}{StochKit2R ensemble (from ssa or tauLeaping) output data}

\item{outputDir}{Character string with path to output directory. If specified output directory does not exist, it will be created. If output directory or file already exists, use \code{force=TRUE} to overwrite.}

\item{force}{Force overwriting of existing data. Warning: force=TRUE will delete all existing content in the output directory.}

\item{indexStart}{Starting index. Default =1 for R data. Set to 0 for compatibility with StochKit2 C++ version.}
}
\description{
\code{writeEnsembleData} Writes StochKit2R ensemble output data to files.
}
\examples{
\dontrun{
#'#example using included dimer_decay.xml file
#run 100 simulations for 10 time units, keeping output at 20 time intervals
#store model file name in a variable first
model <- system.file("dimer_decay.xml",package="StochKit2R")
out <- ssa(modelFile=model,time=10,realizations=100,intervals=20)
#creates ex_out directory and writes output
writeEnsembleData(data=out,outputDir="ex_out")
}
}
