% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutputDirs.R
\name{createOutputDirs}
\alias{createOutputDirs}
\title{Create Output Directories for ssa and tauLeaping}
\usage{
createOutputDirs(outputDir, subdirs, force = FALSE)
}
\arguments{
\item{outputDir}{Character string with path to output directory.}

\item{subdirs}{Character vector containing names of subdirectories to create under outputDir}

\item{force}{delete and overwrite existing output directories, if they exists}
}
\description{
\code{createOutputDirs} Called by writeOutput create output directory
}
