% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotHistogram.R
\name{plotHistogram}
\alias{plotHistogram}
\title{Plot StochKit2R simulation histogram output data}
\usage{
plotHistogram(data, species, timeIndex = NULL)
}
\arguments{
\item{data}{ensemble output from ssa or tauLeaping (histograms object must exist, i.e. ensemble must have been run with keepHistograms=TRUE).}

\item{species}{The species name or index. For numeric indexes, the first species is index 1.}

\item{timeIndex}{The time index (the initial condition is index 1, the index of the end time is equal to the number of output intervals+1). By default =NULL the last index is used.}
}
\value{
The ggplot object
}
\description{
\code{plotHistogram} Plots histogram of data stored in StochKit2R output object or histogram output file.
}
\examples{
\dontrun{
#example using included dimer_decay.xml file
model <- system.file("dimer_decay.xml", package="StochKit2R")
#output written to ex_out directory (created in current working directory)
out <- ssa(model, time=10, realizations=100, intervals=20, keepHistograms=TRUE)
#plot the histogram for species "S2" at time point 5 (t=2.0)
plotHistogram(out, species="S2", timeIndex=5)
}
}
