% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotStats.R
\name{plotStats}
\alias{plotStats}
\title{Plot StochKit2R simulation statistics output data}
\usage{
plotStats(data, species = NULL)
}
\arguments{
\item{data}{ensemble output from ssa or tauLeaping (stats object must exist, i.e. ensemble must NOT have been run with noStats=TRUE).}

\item{species}{A character vector of species names or a numeric vector of species indexes of the species that will be plotted. For numeric indexes, the first species is index 1. By default =NULL, all species are plotted.}
}
\value{
The ggplot object
}
\description{
\code{plotStats} Plots means and mean +/- one standard deviation of populations specified in \code{species}
}
\examples{
\dontrun{
#example using included dimer_decay.xml file
model <- system.file("dimer_decay.xml", package="StochKit2R")
#output written to ex_out directory (created in current working directory)
out <- ssa(model, time=10, realizations=100, intervals=20)
#plot the data for all species
plotStats(out)
#plot the data for species S2 and S3
plotStats(out, species=c("S2", "S3"))
}
}
