% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrajectories.R
\name{plotTrajectories}
\alias{plotTrajectories}
\title{Plot StochKit2R simulation trajectory output data}
\usage{
plotTrajectories(data, trajectoryIndexes = NULL, species = NULL)
}
\arguments{
\item{data}{ensemble output from ssa or tauLeaping (trajectories object must exist, i.e. ensemble must have been run with keepTrajectories=TRUE).}

\item{trajectoryIndexes}{Integer vector containing indexes for which simulations will be plotted. By default (=NULL) all trajectories are plotted.}

\item{species}{A character vector of species names or a numeric vector of species indexes of the species that will be plotted. For numeric indexes, the first species is index 1. By default =NULL, all species are plotted.}
}
\value{
The ggplot object
}
\description{
\code{plotTrajectories} Plots trajectories \code{trajectoryIndexes} of populations specified in \code{species} in the StochKit2R trajectories element of \code{data}
}
\examples{
\dontrun{
#example using included dimer_decay.xml file
model <- system.file("dimer_decay.xml",package="StochKit2R")
#output written to ex_out directory (created in current working directory)
out <- ssa(model,time=10,realizations=100,intervals=20,keepTrajectories=TRUE)
#plot the data for species S2 and S3 for trajectories 2,3,4 and 5
plotTrajectories(out,trajectoryIndexes=2:5,species=c("S2","S3"))
}
}
